requirements.txt
Faker==5.3.0
psycopg2==2.8.6
pytest==6.2.1
pytest-xdist==2.2.0
PyYAML==5.3.1
requests==2.25.1
retrying==1.3.3
urllib3==1.26.2
allure-pytest==2.8.29
pytest-rerunfailures==9.1.1